Учтены замечания
