+++ src/main/java/ru/mail/polis/homework/analyzer/IsSpam.java
@@ -2,11 +2,13 @@ package ru.mail.polis.homework.analyzer;
 
 public class IsSpam implements TextAnalyzer {
 
-    String[] textSpam;
+    private int priority = 0;
+    private String[] textSpam;
     public IsSpam(String[] spam) {
         textSpam = spam;
     }
 
+    @Override
     public FilterType analyze(String arg) {
         for (String s : textSpam) {
             if (arg.contains(s)) {
@@ -16,8 +18,13 @@ public class IsSpam implements TextAnalyzer {
         return FilterType.GOOD;
     }
 
-    private int priority = 0;
+    @Override
     public int getPriority() {
         return priority;
+    }
+
+    @Override
+    public void setPriority(int newPriority) {
+        this.priority = newPriority;
     }
 }
